GCW-3432 Move MainRouter to components folder
