+++ src/App.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import { IonApp } from "@ionic/react";
 import { IonReactRouter } from "@ionic/react-router";
 import AuthContext from "./context/AuthContext";
-import MainRouter from "./MainRouter/MainRouter";
+import MainRouter from "./components/MainRouter/MainRouter";
 
 /* Core CSS required for Ionic components to work properly */
 import "@ionic/react/css/core.css";
