+++ src/components/MainRouter/MainRouter.test.tsx
@@ -3,7 +3,7 @@ import { render, screen } from "@testing-library/react";
 import MainRouter from "./MainRouter";
 import { createMemoryHistory } from "history";
 import { MemoryRouter, Router } from "react-router";
-import AuthContext from "../context/AuthContext";
+import AuthContext from "../../context/AuthContext";
 
 test("should render an IonRouterOutlet", () => {
   render(<MainRouter />, { wrapper: MemoryRouter });
