+++ src/components/MainRouter/MainRouter.tsx
@@ -1,9 +1,9 @@
 import { IonRouterOutlet } from "@ionic/react";
 import React, { useContext } from "react";
 import { Redirect, Route } from "react-router";
-import AuthContext from "../context/AuthContext";
-import Home from "../pages/Home";
-import Login from "../pages/Login/Login";
+import AuthContext from "../../context/AuthContext";
+import Home from "../../pages/Home";
+import Login from "../../pages/Login/Login";
 
 const MainRouter: React.FC = () => {
   const { isAuthenticated } = useContext(AuthContext);
